refactor: change answer status code
