+++ src/auth/auth.guard.ts
@@ -7,7 +7,7 @@ import {
 import { Reflector } from '@nestjs/core';
 import { JwtService } from '@nestjs/jwt';
 import { Request } from 'express';
-import { IS_PUBLIC_KEY } from './custom-public-decorator';
+import { IS_PUBLIC_KEY } from './decorators/custom-public.decorator';
 import { SECRET } from 'src/utils/constants';
 
 @Injectable()
